+++ lib/colibri2_mappings.default.ml
@@ -1065,4 +1065,6 @@ module Fresh = struct
   end
 end
 
+let is_available = true
+
 include Fresh.Make ()